Refactor ServiceFlowPage component and useServiceStore
