+++ GUI/src/types/service-flow.ts
@@ -41,6 +41,7 @@ export type NodeDataProps = {
   linkText?: string;
   fileName?: string;
   fileContent?: string;
+  signOption?: string;
 };
 
 export const initialNodes: Node[] = [
